+++ apps/modernization-ui/src/apps/page-builder/page/management/edit/add-question/modal/AddQuestionModal.scss
@@ -1,5 +1,6 @@
 .add-question-modal {
     height: 85vh;
+    min-width: 85%;
     overflow: hidden;
     .usa-modal__content {
         height: 100%;
