+++ apps/modernization-ui/src/apps/page-builder/page/management/edit/add-question/search/table/expanded-question.module.scss
@@ -12,7 +12,6 @@
         padding: 1.5rem;
         flex-wrap: wrap;
         gap: 1.5rem;
-        justify-content: space-between;
         align-items: flex-start;
         flex-shrink: 0;
         border-radius: 0.3125rem;
