+++ apps/modernization-ui/src/apps/page-builder/page/management/edit/add-question/search/table/question-search-table.module.scss
@@ -29,4 +29,10 @@
     .expandButton {
         cursor: pointer;
     }
+
+    thead tr th {
+        &:nth-of-type(6) {
+            width: 4rem;
+        }
+    }
 }
